+++ src/rpc/client.ts
@@ -188,16 +188,13 @@ export class Client extends EventEmitter {
       }
     }
 
-    if (activity.type) {
+    if (activity.type !== undefined) {
       cleaned.type = activity.type;
-      if ([ActivityType.Playing, ActivityType.Listening, ActivityType.Watching, ActivityType.Competing].includes(cleaned.type)) {
-        cleaned.type = ActivityType.Playing;
-      }
     }
 
     await this.#request("SET_ACTIVITY", {
       pid: pid ?? process ? process.pid ?? 0 : 0,
-      activity,
+      activity: cleaned,
     });
   }